Add ability to download committee records excel file
The BillExcelFileDownloader class downloads an Excel file from the Committee Votes page. Playwright library is used to automate the browser, navigate to the Committee Votes URL, locate the download button, start downloading the Excel file, and save it in the user's download folder. The suggested filename is manipulated to replace .xls extension with .html since contents of .xls file is all html if viewed in text editor.